decomission file_service
