+++ core/src/repo/metadata_repo.rs
@@ -1,10 +1,12 @@
+use uuid::Uuid;
+
+use lockbook_models::file_metadata::FileMetadata;
+
 use crate::model::errors::core_err_unexpected;
 use crate::model::repo::RepoSource;
 use crate::model::state::Config;
 use crate::repo::local_storage;
 use crate::CoreError;
-use lockbook_models::file_metadata::FileMetadata;
-use uuid::Uuid;
 
 const NAMESPACE_LOCAL: &str = "changed_local_metadata";
 const NAMESPACE_BASE: &str = "all_base_metadata";
@@ -63,11 +65,12 @@ pub fn delete_all(config: &Config, source: RepoSource) -> Result<(), CoreError> 
 
 #[cfg(test)]
 mod unit_tests {
+    use uuid::Uuid;
+
     use crate::model::repo::RepoSource;
     use crate::model::state::temp_config;
     use crate::repo::metadata_repo;
     use crate::service::test_utils;
-    use uuid::Uuid;
 
     #[test]
     fn get() {
