+++ core/src/service/db_state_service.rs
@@ -1,7 +1,8 @@
+use serde::Serialize;
+
 use crate::model::state::Config;
 use crate::repo::{account_repo, db_version_repo};
 use crate::CoreError;
-use serde::Serialize;
 
 pub fn get_code_version() -> &'static str {
     env!("CARGO_PKG_VERSION")
