+++ core/src/service/file_compression_service.rs
@@ -1,7 +1,8 @@
+use std::io::{Read, Write};
+
 use flate2::read::ZlibDecoder;
 use flate2::write::ZlibEncoder;
 use flate2::Compression;
-use std::io::{Read, Write};
 
 use crate::CoreError;
